# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PRAMANA
  description: Proof-carrying, multilingual assistant for Ayurvedic IP, ABS and drug-regulatory
    questions. Informational, not legal advice.
  version: 0.1.0
paths:
...
  /v1/documents:
    get:
      tags:
      - documents
      summary: List Documents
      operationId: list_documents_v1_documents_get
      parameters:
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Jurisdiction'
          - type: 'null'
          title: Jurisdiction
      - name: doc_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Doc Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentSummary'
                title: Response List Documents V1 Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/documents/{doc_id}/pdf:
    get:
      tags:
      - documents
      summary: Get Document Pdf
      operationId: get_document_pdf_v1_documents__doc_id__pdf_get
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      - name: corpus_version
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Corpus Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/spans/{evidence_id}:
    get:
      tags:
      - documents
      summary: Get Span
      operationId: get_span_v1_spans__evidence_id__get
      parameters:
      - name: evidence_id
        in: path
        required: true
        schema:
          type: string
          title: Evidence Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceSpan'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/receipts/{receipt_id}:
    get:
      tags:
      - receipts
      summary: Get Receipt
      operationId: get_receipt_v1_receipts__receipt_id__get
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Receipt'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/receipts/{receipt_id}/verify:
    post:
      tags:
      - receipts
      summary: Verify Receipt
      operationId: verify_receipt_v1_receipts__receipt_id__verify_post
      parameters:
      - name: receipt_id
        in: path
        required: true
        schema:
          type: string
          title: Receipt Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/escalations:
    post:
      tags:
      - escalations
      summary: Create Escalation
      operationId: create_escalation_v1_escalations_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EscalationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscalationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - escalations
      summary: List Escalations
      operationId: list_escalations_v1_escalations_get
      parameters:
      - name: X-Demo-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Demo-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Escalations V1 Escalations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DocType:
      type: string
      enum:
      - statute
      - rule
      - regulation
      - treaty
      - notification
      - case
      - guideline
      - manual
      title: DocType
    DocumentSummary:
      properties:
        id:
          type: string
          title: Id
        short_key:
          type: string
          title: Short Key
        title:
          type: string
          title: Title
        doc_type:
          $ref: '#/components/schemas/DocType'
        jurisdiction:
          $ref: '#/components/schemas/Jurisdiction'
        issuer:
          anyOf:
          - type: string
          - type: 'null'
          title: Issuer
        source_url:
          type: string
          title: Source Url
        language:
          type: string
          title: Language
        in_force_from:
          type: string
          format: date
          title: In Force From
        in_force_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: In Force To
      additionalProperties: false
      type: object
      required:
      - id
      - short_key
      - title
      - doc_type
      - jurisdiction
      - source_url
      - language
      - in_force_from
      title: DocumentSummary
...
    EscalationRequest:
      properties:
        request_id:
          type: string
          title: Request Id
        contact:
          anyOf:
          - type: string
          - type: 'null'
          title: Contact
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - request_id
      title: EscalationRequest
    EscalationResponse:
      properties:
        ticket_id:
          type: string
          title: Ticket Id
        status:
          type: string
          title: Status
      additionalProperties: false
      type: object
      required:
      - ticket_id
      - status
      title: EscalationResponse
...
    EvidenceSpan:
      properties:
        id:
          type: string
          title: Id
        doc_id:
          type: string
          title: Doc Id
        doc_title:
          type: string
          title: Doc Title
        doc_type:
          $ref: '#/components/schemas/DocType'
        jurisdiction:
          type: string
          enum:
          - IN
          - INTL
          title: Jurisdiction
        citation_label:
          type: string
          title: Citation Label
        section_key:
          type: string
          title: Section Key
        section_path:
          items:
            type: string
          type: array
          title: Section Path
        page:
          type: integer
          title: Page
        page_end:
          type: integer
          title: Page End
        char_start:
          type: integer
          title: Char Start
        char_end:
          type: integer
          title: Char End
        text:
          type: string
          title: Text
        sha256:
          type: string
          title: Sha256
        effective_from:
          type: string
          format: date
          title: Effective From
        effective_to:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Effective To
        corpus_version:
          type: string
          title: Corpus Version
        source_url:
          type: string
          title: Source Url
        pdf_url:
          type: string
          title: Pdf Url
        highlights:
          items:
            $ref: '#/components/schemas/Highlight'
          type: array
          title: Highlights
      additionalProperties: false
      type: object
      required:
      - id
      - doc_id
      - doc_title
      - doc_type
      - jurisdiction
      - citation_label
      - section_key
      - section_path
      - page
      - page_end
      - char_start
      - char_end
      - text
      - sha256
      - effective_from
      - corpus_version
      - source_url
      - pdf_url
      title: EvidenceSpan
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Highlight:
      properties:
        page:
          type: integer
          title: Page
        page_width:
          type: number
          title: Page Width
        page_height:
          type: number
          title: Page Height
        rects:
          items:
            items:
              type: number
            type: array
          type: array
          title: Rects
          description: Each rect is [x0, y0, x1, y1] in PDF points, origin top-left
            (PyMuPDF).
      additionalProperties: false
      type: object
      required:
      - page
      - page_width
      - page_height
      - rects
      title: Highlight
...
    Jurisdiction:
      type: string
      enum:
      - IN
      - INTL
      - BOTH
      title: Jurisdiction
...
    ModelIds:
      properties:
        llm:
          type: string
          title: Llm
        embed:
          type: string
          title: Embed
        nli:
          type: string
          title: Nli
      additionalProperties: false
      type: object
      required:
      - llm
      - embed
      - nli
      title: ModelIds
...
    Receipt:
      properties:
        id:
          type: string
          title: Id
        request_id:
          type: string
          title: Request Id
        corpus_version:
          type: string
          title: Corpus Version
        query_hash:
          type: string
          title: Query Hash
        chunk_hashes:
          items:
            type: string
          type: array
          title: Chunk Hashes
        model_ids:
          $ref: '#/components/schemas/ModelIds'
        prompt_version:
          type: string
          title: Prompt Version
        prev_hash:
          type: string
          title: Prev Hash
        entry_hash:
          type: string
          title: Entry Hash
        created_at:
          type: string
          format: date-time
          title: Created At
      additionalProperties: false
      type: object
      required:
      - id
      - request_id
      - corpus_version
      - query_hash
      - chunk_hashes
      - model_ids
      - prompt_version
      - prev_hash
      - entry_hash
      - created_at
      title: Receipt
...
    SpanVerification:
      properties:
        evidence_id:
          type: string
          title: Evidence Id
        sha256:
          type: string
          title: Sha256
        in_corpus:
          type: boolean
          title: In Corpus
        merkle_proof_valid:
          type: boolean
          title: Merkle Proof Valid
      additionalProperties: false
      type: object
      required:
      - evidence_id
      - sha256
      - in_corpus
      - merkle_proof_valid
      title: SpanVerification
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerifyResult:
      properties:
        chain_valid:
          type: boolean
          title: Chain Valid
        corpus_root:
          type: string
          title: Corpus Root
        spans:
          items:
            $ref: '#/components/schemas/SpanVerification'
          type: array
          title: Spans
      additionalProperties: false
      type: object
      required:
      - chain_valid
      - corpus_root
      title: VerifyResult
...